nuk-line-bot/heroku/requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
requests==2.18.4
urllib3==1.22
psycopg2==2.7.3.2
line-bot-sdk
flask
